// samples/web-csr/dressca-frontend/admin/src/generated/api-client/models/catalog-item-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dressca Web API
 * Dressca の Web API 仕様
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * カタログアイテムのレスポンスデータを表します。             
 * @export
 * @interface CatalogItemResponse
 */
export interface CatalogItemResponse {
    /**
     * 説明を取得または設定します。             
     * @type {string}
     * @memberof CatalogItemResponse
     */
    'description': string;
    /**
     * 単価を取得または設定します。             
     * @type {number}
     * @memberof CatalogItemResponse
     */
    'price': number;
    /**
     * カタログカテゴリ Id を取得または設定します。             
     * @type {number}
     * @memberof CatalogItemResponse
     */
    'catalogCategoryId': number;
    /**
     * カタログブランド Id を取得または設定します。             
     * @type {number}
     * @memberof CatalogItemResponse
     */
    'catalogBrandId': number;
    /**
     * カタログアイテム Id を取得または設定します。             
     * @type {number}
     * @memberof CatalogItemResponse
     */
    'id': number;
    /**
     * 商品名を取得または設定します。             
     * @type {string}
     * @memberof CatalogItemResponse
     */
    'name': string;
    /**
     * 商品コードを取得または設定します。             
     * @type {string}
     * @memberof CatalogItemResponse
     */
    'productCode': string;
    /**
     * アセットコードの一覧を取得または設定します。             
     * @type {Array<string>}
     * @memberof CatalogItemResponse
     */
    'assetCodes'?: Array<string>;
    /**
     * 行バージョンを取得または設定します。
     * @type {string}
     * @memberof CatalogItemResponse
     */
    'rowVersion': string;
}